+++ examples/refresh-token-sample/platforms/ios/www/lib/auth0-angular/Gruntfile.js
@@ -22,6 +22,19 @@ module.exports = function (grunt) {
   ];
 
   grunt.initConfig({
+    pkg: grunt.file.readJSON('package.json'),
+    meta: {
+      banner: [
+                '/**',
+                ' * <%= pkg.description %>',
+                ' * @version v<%= pkg.version %> - <%= grunt.template.today("yyyy-mm-dd") %>',
+                ' * @link <%= pkg.homepage %>',
+                ' * @author <%= pkg.author %>',
+                ' * @license MIT License, http://www.opensource.org/licenses/MIT',
+                ' */',
+                ''
+              ].join('\n')
+    },
     clean: [
       'build/'
     ],
@@ -48,6 +61,9 @@ module.exports = function (grunt) {
     },
 
     concat: {
+      options: {
+        banner: '<%= meta.banner %>'
+      },
       dist: { dest: 'build/auth0-angular.js', src: ['build/auth0-angular.js'] }
     },
 
@@ -76,6 +92,7 @@ module.exports = function (grunt) {
           'examples/widget/scripts/auth0-angular.js':                     'build/auth0-angular.js',
           'examples/sso/scripts/auth0-angular.js':                        'build/auth0-angular.js',
           'examples/widget-redirect/scripts/auth0-angular.js':            'build/auth0-angular.js',
+          'examples/redirect/scripts/auth0-angular.js':            'build/auth0-angular.js',
           'examples/ui-router/scripts/auth0-angular.js':                  'build/auth0-angular.js',
           'examples/refresh-token/auth0-angular.js':                  'build/auth0-angular.js',
           'examples/requirejs/scripts/auth0-angular.js':                  'build/auth0-angular.js'