+++ examples/refresh-token-sample/www/lib/auth0-angular/examples/custom-signup/client/scripts/auth0-angular.js
@@ -1,3 +1,10 @@
+/**
+ * Angular SDK to use with Auth0
+ * @version v3.0.5 - 2014-12-02
+ * @link https://auth0.com
+ * @author Martin Gontovnikas
+ * @license MIT License, http://www.opensource.org/licenses/MIT
+ */
 (function () {
   angular.module('auth0', [
     'auth0.service',
@@ -12,6 +19,12 @@
     var Utils = {
         capitalize: function (string) {
           return string ? string.charAt(0).toUpperCase() + string.substring(1).toLowerCase() : null;
+        },
+        fnName: function (fun) {
+          var ret = fun.toString();
+          ret = ret.substr('function '.length);
+          ret = ret.substr(0, ret.indexOf('('));
+          return ret ? ret.trim() : ret;
         }
       };
     angular.extend(this, Utils);
@@ -116,7 +129,7 @@
             }
           },
           'Auth0Lock': {
-            signin: 'showSignin',
+            signin: 'show',
             signup: 'showSignup',
             reset: 'showReset',
             library: function () {
@@ -152,10 +165,10 @@
         if (!Constructor && typeof Auth0 !== 'undefined') {
           Constructor = Auth0;
         }
-        if (Constructor.name === 'Auth0Widget') {
+        if (authUtilsProvider.fnName(Constructor) === 'Auth0Widget') {
           throw new Error('Auth0Widget is not supported with this ' + ' version of auth0-angular anymore. Please try with an older one');
         }
-        if (Constructor.name === 'Auth0Lock') {
+        if (authUtilsProvider.fnName(Constructor) === 'Auth0Lock') {
           this.auth0lib = new Constructor(this.clientID, domain, angular.extend(defaultOptions, options));
           this.auth0js = this.auth0lib.getClient();
           this.isLock = true;
@@ -254,7 +267,7 @@
           if (config.loginUrl) {
             $rootScope.$on('$routeChangeStart', function (e, nextRoute) {
               if (nextRoute.$$route && nextRoute.$$route.requiresLogin) {
-                if (!auth.isAuthenticated) {
+                if (!auth.isAuthenticated && !auth.refreshTokenPromise) {
                   $location.path(config.loginUrl);
                 }
               }
@@ -263,7 +276,7 @@
           if (config.loginState) {
             $rootScope.$on('$stateChangeStart', function (e, to) {
               if (to.data && to.data.requiresLogin) {
-                if (!auth.isAuthenticated) {
+                if (!auth.isAuthenticated && !auth.refreshTokenPromise) {
                   e.preventDefault();
                   $injector.get('$state').go(config.loginState);
                 }
@@ -292,9 +305,12 @@
           };
           auth.refreshIdToken = function (refresh_token) {
             var refreshTokenAsync = authUtils.promisify(config.auth0js.refreshToken, config.auth0js);
-            return refreshTokenAsync(refresh_token || auth.refreshToken).then(function (delegationResult) {
+            auth.refreshTokenPromise = refreshTokenAsync(refresh_token || auth.refreshToken).then(function (delegationResult) {
               return delegationResult.id_token;
+            })['finally'](function () {
+              auth.refreshTokenPromise = null;
             });
+            return auth.refreshTokenPromise;
           };
           auth.renewIdToken = function (id_token) {
             var renewIdTokenAsync = authUtils.promisify(config.auth0js.renewIdToken, config.auth0js);
@@ -366,7 +382,7 @@
             callHandler('logout');
           };
           auth.authenticate = function (profile, idToken, accessToken, state, refreshToken) {
-            onSigninOk(idToken, accessToken, state, refreshToken, profile, true);
+            return onSigninOk(idToken, accessToken, state, refreshToken, profile, true);
           };
           auth.getProfile = function (idToken) {
             var getProfilePromisify = authUtils.promisify(config.auth0lib.getProfile, config.auth0lib);