+++ examples/refresh-token-sample/www/lib/auth0-angular/examples/widget-redirect/scripts/myApp.js
@@ -27,7 +27,7 @@ myApp.config(function ($routeProvider, authProvider, $httpProvider, jwtIntercept
     loginUrl: '/login'
   });
 
-  authProvider.on('loginSuccess', function($location, profilePromise, idToken) {
+  authProvider.on('loginSuccess', function($location, profilePromise, idToken, store) {
     console.log("Login Success");
     profilePromise.then(function(profile) {
       store.set('profile', profile);
@@ -57,15 +57,18 @@ myApp.config(function ($routeProvider, authProvider, $httpProvider, jwtIntercept
   // NOTE: in case you are calling APIs which expect a token signed with a different secret, you might
   // want to check the delegation-token example
   $httpProvider.interceptors.push('jwtInterceptor');
-}).run(function($rootScope, auth, store) {
+}).run(function($rootScope, auth, store, jwtHelper, $location) {
   $rootScope.$on('$locationChangeStart', function() {
     if (!auth.isAuthenticated) {
       var token = store.get('token');
       if (token) {
-        auth.authenticate(store.get('profile'), token);
+        if (!jwtHelper.isTokenExpired(token)) {
+          auth.authenticate(store.get('profile'), token);
+        } else {
+          $location.path('/login');
+        }
       }
     }
 
   });
 });
-