+++ examples/refresh-token-sample/platforms/ios/www/lib/auth0-angular/CHANGELOG.md
@@ -1,5 +1,25 @@
 # 3.0.X
 
+## 3.0.6
+* Added versioning info to built files
+
+## 3.0.5
+* Fixed bug with IE and `Function.name`
+* Added `redirect` mode example for `Auth0.js`
+
+## 3.0.4
+* Now the signin by default shows the signup and reset buttons
+
+## 3.0.3
+* Tutorial fixes for when the page is refreshed and the JWT is expired
+* Added handling for the case when JWT is expired, page is refreshed and a RefreshToken is used to refresh the page.
+
+## 3.0.2
+* Minor fixes
+
+## 3.0.1
+* Minor fixes
+
 ## 3.0.0
 * Magic is removed. `auth0-angular` is now much smaller and with less magic
 * We recommend the usage of `angular-storage` and `angular-jwt` for storing information and token handling